feat: bump (#566)
* feat: bump

* fixed tests

+++ crates/sui-prover/tests/inputs/dynamic_field/exists_.ok.move
@@ -9,14 +9,14 @@ public struct Foo has key {
 }
 
 fun foo(x: &mut Foo) {
-    dynamic_field::add<u64, u8>(&mut x.id, 10, 0);
+    dynamic_field::add<u64, u64>(&mut x.id, 10u64, 0);
 }
 
 #[spec(prove)]
 fun foo_spec(x: &mut Foo) {
-    requires(!dynamic_field::exists_with_type<u64, u8>(&x.id, 10));
+    requires(!dynamic_field::exists_with_type<u64, u64>(&x.id, 10u64));
     foo(x);
-    ensures(dynamic_field::exists_(&x.id, 10));
-    ensures(dynamic_field::exists_with_type<u64, u8>(&x.id, 10));
-    ensures(dynamic_field::borrow<u64, u8>(&x.id, 10) == 0);
+    ensures(dynamic_field::exists_(&x.id, 10u64));
+    ensures(dynamic_field::exists_with_type<u64, u64>(&x.id, 10u64));
+    ensures(dynamic_field::borrow<u64, u64>(&x.id, 10) == 0);
 }
